Fix failure because of race conditions for `update-resolv-conf.service` (#1085)
* Fix failure because of race conditions for update-resolv-conf.service

* adapt tests

* modify update-resolv-conf.sh

* edit tests

* Make exit status match sysexits' EX_CONFIG

+++ pkg/webhook/controlplane/ensurer.go
@@ -239,6 +239,8 @@ StartLimitIntervalSec=0
 [Service]
 Type=oneshot
 ExecStart=/opt/bin/update-resolv-conf.sh
+RestartForceExitStatus=78
+RestartSec=15
 `
 	)
 
@@ -283,6 +285,10 @@ is_systemd_resolved_system()
       return 1
     fi
 }
+
+if ! grep -Eq "^nameserver\s+" /run/systemd/resolve/resolv.conf; then
+  exit 78
+fi
 
 rm -f "$tmp"
 if is_systemd_resolved_system; then
